requirements.txt
fastapi
uvicorn
sentence-transformers
psycopg2-binary
python-dotenv
pandas
scikit-learn
